backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.1
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pymongo==4.3.3